pyproject.toml
[project]
name = "SICrMLB-Models"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "artumont", email = "197291181+artumont@users.noreply.github.com"},
]
dependencies = ["pillow>=12.0.0", "numpy>=2.3.4", "ultralytics>=8.3.217", "pydantic>=2.12.3"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
